// src/types/champion.ts
// Generated by https://quicktype.io

export interface Champion {
	id: number
	key: string
	name: string
	title: string
	fullName: string
	icon: string
	resource: string
	attackType: string
	adaptiveType: string
	stats: { [key: string]: Stat }
	roles: string[]
	attributeRatings: AttributeRatings
	abilities: Abilities
	releaseDate: string
	releasePatch: string
	patchLastChanged: string
	price: Price
	lore: string
	faction: string
	skins: Skin[]
}

export interface Abilities {
	P: E[]
	Q: E[]
	W: E[]
	E: E[]
	R: E[]
}

export interface E {
	name: string
	icon: string
	effects: Effect[]
	cost: null
	cooldown: Cooldown
	targeting: string
	affects: string
	spellshieldable: null | string
	resource: null
	damageType: null | string
	spellEffects: null | string
	projectile: null | string
	onHitEffects: null
	occurrence: null
	notes: string
	blurb: string
	missileSpeed: null
	rechargeRate: null
	collisionRadius: null
	tetherRadius: null
	onTargetCdStatic: null
	innerRadius: null
	speed: null | string
	width: null | string
	angle: null
	castTime: null | string
	effectRadius: null | string
	targetRange: null | string
}

export interface Cooldown {
	modifiers: Modifier[]
	affectedByCdr: boolean
}

export interface Modifier {
	values: number[]
	units: Unit[]
}

export enum Unit {
	Ad = "% AD",
	Empty = "",
	Unit = "%",
}

export interface Effect {
	description: string
	leveling: Leveling[]
}

export interface Leveling {
	attribute: string
	modifiers: Modifier[]
}

export interface AttributeRatings {
	damage: number
	toughness: number
	control: number
	mobility: number
	utility: number
	abilityReliance: number
	attack: number
	defense: number
	magic: number
	difficulty: number
}

export interface Price {
	blueEssence: number
	rp: number
	saleRp: number
}

export interface Skin {
	name: string
	id: number
	isBase: boolean
	availability: Availability
	formatName: string
	lootEligible: boolean
	cost: number | string
	sale: number
	distribution: null | string
	rarity: RarityEnum
	chromas: Chroma[]
	lore: null | string
	release: string
	set: string[]
	splashPath: string
	uncenteredSplashPath: string
	tilePath: string
	loadScreenPath: string
	loadScreenVintagePath: null | string
	newEffects: boolean
	newAnimations: boolean
	newRecall: boolean
	newVoice: boolean
	newQuotes: boolean
	voiceActor: VoiceActor[]
	splashArtist: string[]
}

export enum Availability {
	Available = "Available",
	Limited = "Limited",
	Rare = "Rare",
}

export interface Chroma {
	name: string
	id: number
	chromaPath: string
	colors: string[]
	descriptions: Description[]
	rarities: RarityElement[]
}

export interface Description {
	description: null | string
	region: Region | null
}

export enum Region {
	Riot = "riot",
	Tencent = "TENCENT",
}

export interface RarityElement {
	rarity: number | null
	region: Region | null
}

export enum RarityEnum {
	Epic = "Epic",
	Mythic = "Mythic",
	NoRarity = "NoRarity",
}

export enum VoiceActor {
	RamonTikaram = "Ramon Tikaram",
}

export interface Stat {
	flat: number
	percent: number
	perLevel: number
	percentPerLevel: number
}